"cache功能引入和点赞功能的缓存增删改查"

+++ internal/cache/favorite.go
@@ -6,10 +6,10 @@ import (
 	"time"
 )
 
-// 缓存的crud
+// 缓存的favorite-crud
 
 // Favor userId给videoId点赞
-func (r *Redis) Favor(userId uint, videoId uint) (bool, error) {
+func (r *Redis) Favor(userId int64, videoId int64) (bool, error) {
 	key := util.UserFavorKey(userId)
 	err := r.redis.SetBit(key, int64(videoId), 1).Err()
 	if err != nil {
@@ -24,7 +24,7 @@ func (r *Redis) Favor(userId uint, videoId uint) (bool, error) {
 }
 
 // CancelFavor userId给videoId取消点赞
-func (r *Redis) CancelFavor(userId uint, videoId uint) (bool, error) {
+func (r *Redis) CancelFavor(userId int64, videoId int64) (bool, error) {
 	key := util.UserFavorKey(userId)
 	err := r.redis.SetBit(key, int64(videoId), 0).Err()
 	if err != nil {
@@ -39,7 +39,7 @@ func (r *Redis) CancelFavor(userId uint, videoId uint) (bool, error) {
 }
 
 // QueryFavorCnt 查询videoId的点赞数量
-func (r *Redis) QueryFavorCnt(videoId uint) (uint, error) {
+func (r *Redis) QueryFavorCnt(videoId int64) (int64, error) {
 	key := util.VideoFavoredCntKey(videoId)
 	result, err := r.redis.Get(key).Result()
 	if err != nil {
@@ -49,11 +49,11 @@ func (r *Redis) QueryFavorCnt(videoId uint) (uint, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint(cnt), err
+	return int64(cnt), err
 }
 
 // IncrFavorCnt 增加userId的点赞数量
-func (r *Redis) IncrFavorCnt(userId uint) (uint, error) {
+func (r *Redis) IncrFavorCnt(userId int64) (int64, error) {
 	key := util.UserFavorCntKey(userId)
 	//expire会更新key的过期时间，并且如果不存在的话 ‘不会’ 创建这个key
 	err := r.redis.Expire(key, time.Hour).Err()
@@ -64,11 +64,11 @@ func (r *Redis) IncrFavorCnt(userId uint) (uint, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint(cnt), nil
+	return int64(cnt), nil
 }
 
 // DecrFavorCnt  减少userId的点赞数量
-func (r *Redis) DecrFavorCnt(userId uint) (uint, error) {
+func (r *Redis) DecrFavorCnt(userId int64) (int64, error) {
 	key := util.UserFavorCntKey(userId)
 	err := r.redis.Expire(key, time.Hour).Err()
 	if err != nil {
@@ -78,11 +78,11 @@ func (r *Redis) DecrFavorCnt(userId uint) (uint, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint(cnt), nil
+	return int64(cnt), nil
 }
 
 // IncrFavoredCnt 增加videoId的被点赞数量
-func (r *Redis) IncrFavoredCnt(videoId uint) (uint, error) {
+func (r *Redis) IncrFavoredCnt(videoId int64) (int64, error) {
 	key := util.VideoFavoredCntKey(videoId)
 	err := r.redis.Expire(key, time.Hour).Err()
 	if err != nil {
@@ -92,11 +92,11 @@ func (r *Redis) IncrFavoredCnt(videoId uint) (uint, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint(cnt), nil
+	return int64(cnt), nil
 }
 
 // DecrFavoredCnt 减少videoId的被点赞数量
-func (r *Redis) DecrFavoredCnt(videoId uint) (uint, error) {
+func (r *Redis) DecrFavoredCnt(videoId int64) (int64, error) {
 	key := util.VideoFavoredCntKey(videoId)
 	err := r.redis.Expire(key, time.Hour).Err()
 	if err != nil {
@@ -106,11 +106,11 @@ func (r *Redis) DecrFavoredCnt(videoId uint) (uint, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint(cnt), nil
+	return int64(cnt), nil
 }
 
 // IsFavor userId是否给videoId点赞
-func (r *Redis) IsFavor(userId uint, videoId uint) (bool, error) {
+func (r *Redis) IsFavor(userId int64, videoId int64) (bool, error) {
 	key := util.UserFavorKey(userId)
 	result, err := r.redis.GetBit(key, int64(videoId)).Result()
 	if err != nil {
@@ -128,8 +128,8 @@ func (r *Redis) DeleteOneCache(key string) (bool, error) {
 	return true, nil
 }
 
-// FavorList 获取userId点赞列表
-func (r *Redis) FavorList(userId uint) ([]uint, error) {
+// FavorList 获取userId点赞列表(userId点赞了什么)
+func (r *Redis) FavorList(userId int64) ([]int64, error) {
 	key := util.UserFavorKey(userId)
 	result, err := r.redis.Get(key).Result()
 	if err != nil {
@@ -140,8 +140,8 @@ func (r *Redis) FavorList(userId uint) ([]uint, error) {
 	return favorList, nil
 }
 
-// FavoredList 获取点赞videoId列表
-func (r *Redis) FavoredList(videoId uint) ([]uint, error) {
+// FavoredList 获取点赞videoId列表(谁点赞了videoId)
+func (r *Redis) FavoredList(videoId int64) ([]int64, error) {
 	key := util.VideoFavoredKey(videoId)
 	result, err := r.redis.Get(key).Result()
 	if err != nil {
